New file for RV controls for VAM (Questionnaires)
This is a list of all VAM NIST controls families (AC-SI) that are applicable to the project and the associated questions for each set of controls. Total number of questions amount to 400+. Questions range from 4-8 per control or twice that or less for a control family (May vary based on applicability). 

RiskVision states that 14 of the controls are at 100% complete, while 7 controls are from 94%-98% near complete.  Context for the degree of what constitutes "near completion or full completion" can vary based on new project requirements, applicability and/or minor updates.